consolidated country-data and contry-list
